Validate Message IDs
:Release Notes:
MSGID values are now validated to ensure they do not contain any
illegal characters, are non-NULL, and non-empty.

:Detailed Notes:
The MSGID field must not contain any spaces or "curly braces" ({}).

All message id's are now validated to enforce this contstraint, and an
error (kPmLogErr_InvalidMsgID) is returned if it is violated.

:Testing Performed:
The change was installed on a device and tested using PmLogCtl to send
messages. For messages containing any combination of the invalid
characters, the following response was displayed:

    Error logging: 0x6C67000F (InvalidMsgID)

Examining the the log file showed a matching entry.

:QA Notes:

:Issues Addressed:
[GF-7642] msgid should be validated for the new APIs ...

Open-webOS-DCO-1.0-Signed-off-by: Hongjeon Ha <hj.ha@lge.com>
Open-webOS-DCO-1.0-Signed-off-by: Keith Derrick <keith.derrick@lge.com>
Change-Id: I33b1d11c8e35a9290c00eaef10c9fb0a01f237b1
Reviewed-on: https://g2g.palm.com/2753
Reviewed-by: Build Verification
Reviewed-by: Gayathri Srinivasan <gayathri.srinivasan@lge.com>
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>