Fix message id truncation issue
:Release Notes:
Message ids are limited to be not more than 32 characters long.

:Detailed Notes:
This functionality is part of MSGID validation. It had a pointer
issue which caused the message ids to be truncated at 3 characters.
This change fixes the issue.

:Testing Performed:
Compiled locally and tested on locally.

Verified in the log file, that all characters of a message id are
properly displayed with a max of 32 characters.

:QA Notes:

:Issues Addressed:
[GF-7642] msgid should be validated for the new APIs

Open-webOS-DCO-1.0-Signed-off-by: Hongjeon Ha <hj.ha@lge.com>

Change-Id: I45ae099d0409fed9d04b655848517612f57a7cb0
Open-webOS-DCO-1.0-Signed-off-by: Ameya Palande <ameya.palande@lge.com>
Reviewed-on: https://g2g.palm.com/2868
Reviewed-by: Build Verification
Reviewed-by: Ameya Palande <ameya.palande@lge.com>
Reviewed-by: Hongjeon Ha <hj.ha@lge.com>
Tested-by: Hongjeon Ha <hj.ha@lge.com>
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Reviewed-by: Gayathri Srinivasan <gayathri.srinivasan@lge.com>