Fix bug for debug messages
:Release Notes:
PmLogDebug() was not record it's log properly

:Detailed Notes:
PmLogDebug() should not have kv-pairs so that should be added
empty curly braces like "{}".
The logic for adding that has some bugs so fixed it.

:Testing Performed:
Compiled locally and tested on local and TV.
Verified in the log file, error text displayed in
/var/log/legacy-log properly.

:QA Notes:

:Issues Addressed:
[GF-13166] Debug messages are not recorded properly

Open-webOS-DCO-1.0-Signed-off-by: Hongjeon Ha <hj.ha@lge.com>

Change-Id: I6d1c1d027a42c51f2a572fee641f6770d14b980e
Reviewed-on: https://g2g.palm.com/3073
Reviewed-by: Build Verification
Reviewed-by: Hongjeon Ha <hj.ha@lge.com>
Tested-by: Hongjeon Ha <hj.ha@lge.com>
Reviewed-by: Keith Derrick <keith.derrick@lge.com>