Remove dependency on upload server
:Release Notes:
Remove dependency on upload server.
Bypass routing of context to separate files in an official flash
build.

:Detailed Notes:
We need to bypass routing of context to separate files.
This will aggregate all the logs in /var/log/messages.
This will also make sure that default level for all contexts
remain at info.
This is for simulating production scenario in a flash official build.
This feature can be removed when we move to production

:Testing Performed:
Compiled locally and tested on a device.

:QA Notes:

:Issues Addressed:
[GF-45673] pmloglib to route all logs to mainstream log file in an
           official flash build

Open-webOS-DCO-1.0-Signed-off-by: Hongjeon Ha <hj.ha@lge.com>

Change-Id: I5f22a9496c206d57dd552dc8c08aa7da0396c13e
Reviewed-on: https://g2g.palm.com/3825
Reviewed-by: Build Verification
Reviewed-by: Hongjeon Ha <hj.ha@lge.com>
Tested-by: Hongjeon Ha <hj.ha@lge.com>
Reviewed-by: Gayathri Srinivasan <gayathri.srinivasan@lge.com>